setup click handler in initializer
